# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema Booking API",
        "description": "Cinema seat booking API with real-time seat locking via WebSocket.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8004",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/bookings/{booking}/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Pay for a booking (mock)",
                "description": "Immediately confirms payment. No real gateway — demo only.",
                "operationId": "ecf900fbbf827f1536e3b5af6de76436",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "example": "credit_card",
                                        "enum": [
                                            "credit_card",
                                            "debit_card",
                                            "online_banking",
                                            "ewallet"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment confirmed"
                    },
                    "422": {
                        "description": "Booking already paid"
                    },
                    "404": {
                        "description": "Booking not found or not yours"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/showtimes/{showtime}/seats": {
            "get": {
                "tags": [
                    "Seats"
                ],
                "summary": "Get seat map for a showtime",
                "description": "Returns all seats with status: available, locked, or booked. is_mine=true if the logged-in user holds the lock.",
                "operationId": "a8fcde6a2a3a5fa0877da2c332bb0fa2",
                "parameters": [
                    {
                        "name": "showtime",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seats retrieved"
                    }
                }
            }
        },
        "/api/showtimes/{showtime}/seats/lock": {
            "post": {
                "tags": [
                    "Seats"
                ],
                "summary": "Lock seats for 10 minutes",
                "description": "First-come-first-serve. Returns 409 if another user already locked the seat.",
                "operationId": "89f68ae86bdb96ec9e6ed6389b9823d4",
                "parameters": [
                    {
                        "name": "showtime",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "seat_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seats locked"
                    },
                    "409": {
                        "description": "Seat already taken"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Seats"
                ],
                "summary": "Release your seat locks",
                "operationId": "c3f595c9716eea8d86817d273ed1ae06",
                "parameters": [
                    {
                        "name": "showtime",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seats released"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}